+++ useScroll/ScrollThumb.js
@@ -29,7 +29,7 @@ const SpotlightAccelerator = new Accelerator();
 let ScrollThumb = forwardRef((props, ref) => {
 	const
 		{cbAlertThumb, focusableScrollbar, onInteractionForScroll, rtl, vertical, ...rest} = props,
-		className = classNames(css.scrollTrack, vertical ? css.vertical : null),
+		className = classNames(css.scrollTrack, vertical ? css.vertical : null, focusableScrollbar ? css.focusableScrollbar : null),
 		ScrollThumbDiv = focusableScrollbar ? Spottable('div') : 'div';
 
 	useEffect (() => {
@@ -38,6 +38,7 @@ let ScrollThumb = forwardRef((props, ref) => {
 
 	useEffect (() => {
 		SpotlightAccelerator.reset();
+		cbAlertThumb(true);
 		return () => {
 			SpotlightAccelerator.reset();
 		};